services/agents/agent_returns/pyproject.toml
[project]
name = "agent_returns"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "google-adk[a2a]>=1.31.1",
    "litellm>=1.83.0",
    "uvicorn>=0.46.0",
]
